requirements.txt
pyfiglet==0.8.post1
PyYAML==6.0
